Fix prev button appearing for first exec
The issue was that the forloop over all the execs was still looping
over inactive execs. So even if they're not appearing in the output,
their sort order (index) still holds.

To fix this, the execs collection is filtered out before sorting to
only include the active ones. This means that checking for whether
they're active is no longer necessary within the iteration.

The issue for the next button collapsing on the first exec is also
fixed by adding an invisible dummy to prevent the next button
from assuming the prev button styling.