Move section margin styles to local section stylings
Instead of having the margin-bottom rule global for all sections,
make it a utility class and add it to each appropriate section
individually.

This is because having the former necesseciates the duplication of
<section>s in the loop, each having a different style based on
mobile or full screen. But this messes up the vertical nav
since now there's two copies of each in the DOM.